+++ deprecated-react-native-prop-types/DeprecatedTransformPropTypes.js
@@ -11,41 +11,10 @@
 
 const PropTypes = require('prop-types');
 
-const TransformMatrixPropType = function (props, propName, componentName) {
-  if (props[propName]) {
-    return new Error(
-      'The transformMatrix style property is deprecated. ' +
-        'Use `transform: [{ matrix: ... }]` instead.',
-    );
-  }
-};
-
-const DecomposedMatrixPropType = function (props, propName, componentName) {
-  if (props[propName]) {
-    return new Error(
-      'The decomposedMatrix style property is deprecated. ' +
-        'Use `transform: [...]` instead.',
-    );
-  }
-};
-
 /**
  * @see facebook/react-native/Libraries/StyleSheet/private/_TransformStyle.js
  */
 const DeprecatedTransformPropTypes = {
-  decomposedMatrix: DecomposedMatrixPropType,
-  rotation: deprecatedPropType(
-    PropTypes.number,
-    'Use the transform prop instead.',
-  ),
-  scaleX: deprecatedPropType(
-    PropTypes.number,
-    'Use the transform prop instead.',
-  ),
-  scaleY: deprecatedPropType(
-    PropTypes.number,
-    'Use the transform prop instead.',
-  ),
   transform: PropTypes.arrayOf(
     PropTypes.oneOfType([
       PropTypes.shape({perspective: PropTypes.number}),
@@ -62,30 +31,6 @@ const DeprecatedTransformPropTypes = {
       PropTypes.shape({translateY: PropTypes.number}),
     ]),
   ),
-  transformMatrix: TransformMatrixPropType,
-  translateX: deprecatedPropType(
-    PropTypes.number,
-    'Use the transform prop instead.',
-  ),
-  translateY: deprecatedPropType(
-    PropTypes.number,
-    'Use the transform prop instead.',
-  ),
 };
-
-/**
- * Adds a deprecation warning when the prop is used.
- */
-function deprecatedPropType(propType, explanation) {
-  return function validate(props, propName, componentName, ...rest) {
-    if (props[propName] !== undefined) {
-      console.warn(
-        `\`${propName}\` supplied to \`${componentName}\` has been deprecated. ${explanation}`,
-      );
-    }
-
-    return propType(props, propName, componentName, ...rest);
-  };
-}
 
 module.exports = DeprecatedTransformPropTypes;