+++ deprecated-react-native-prop-types/CHANGELOG.md
@@ -23,3 +23,7 @@ ImagePropTypes`
   - Added `accessibilityActions`, `accessibilityHint`, `accessibilityLabel`, `accessibilityLanguage`, `accessibilityRole`, `accessibilityState`, `aria-busy`, `aria-checked`, `aria-disabled`, `aria-expanded`, `aria-label`, `aria-labelledby`, `aria-selected`, `dynamicTypeRamp`, `id`, `lineBreakStrategyIOS`, `onAccessibilityAction`, and `role`.
   - Added `onPressIn`, `onPressOut`, `onResponderGrant`, `onResponderMove`, `onResponderRelease`, `onResponderTerminate`, `onResponderTerminationRequest`, `onStartShouldSetResponder`, `onMoveShouldSetResponder`, and `onTextLayout`.
   - Added numeric and named values for `fontWeight` in `style`.
+  - Added string and stylistic values for `fontVariant` in `style`.
+  - Added `userSelect` and `verticalAlign` in `style`.
+- Improvements to `TransformPropTypes`
+  - Removed deprecated props: `decomposedMatrix`, `rotation`, `scaleX`, `scaleY`, `transformMatrix`, `translateX`, `translateY`